backend/app/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.12